+++ api/src/main/kotlin/org/kryptonmc/krypton/api/inventory/item/ItemStack.kt
@@ -1,6 +1,9 @@
 package org.kryptonmc.krypton.api.inventory.item
 
+import org.kryptonmc.krypton.api.inventory.item.meta.ItemMeta
+
 data class ItemStack(
     val type: ItemType,
-    val amount: Int
+    val amount: Int,
+    val meta: ItemMeta? = null
 )