+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.17] - 2021-04-04
+### Added
+- Inventories are now better than they were and exposed properly to the API
+
 ## [0.16.3] - 2021-04-07
 ### Changed
 - Fixed error handling so the disconnect packet varies based on the session's current state